refactor(tui): extract overlay modal and banner rendering component
